backend/requirements.txt
flask
flask-cors
yfinance
transformers
torch
python-dotenv
requests
pandas
numpy
supabase
